Refactor SocialAccountService
Removed type hinting for providerId because it's not integer only

+++ app/Http/Controllers/Auth/LineAccountController.php
@@ -35,7 +35,7 @@ class LineAccountController extends SocialAccountController
         $isAjax = request()->ajax();
 
         try {
-            $message = $accountService->testNotify($this->getProvider(), (int)request('id'), request('message'));
+            $message = $accountService->testNotify($this->getProvider(), request('id'), request('message'));
         } catch (\Exception $e) {
             report($e);
             if ($isAjax) {
