+++ app/Http/Controllers/Auth/TwitterAccountController.php
@@ -48,7 +48,7 @@ class TwitterAccountController extends SocialAccountController
         $isAjax = request()->ajax();
 
         try {
-            $message = $accountService->testNotify($this->getProvider(), (int)request('id'), request('message'));
+            $message = $accountService->testNotify($this->getProvider(), request('id'), request('message'));
         } catch (\Exception $e) {
             if ($isAjax) {
                 return new JsonResponse([
